# docs/swagger.yaml
basePath: /api
definitions:
  model.AdminReq:
    properties:
      name:
        type: string
      password:
        type: string
    required:
    - name
    - password
    type: object
  model.AdminResp:
    properties:
      contact:
        type: string
      name:
        type: string
    type: object
...
host: ralxyz.dev.zjuqsc.com
info:
  contact: {}
  description: This API will be used under staging environment.
  title: Book Management System API
  version: "0.2"
paths:
...
  /login:
    get:
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.AdminResp'
        "401":
          description: Not logged in
          schema:
            type: string
      summary: Check login status
      tags:
      - Login
    post:
      consumes:
      - application/json
      parameters:
      - description: Login information
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/model.AdminReq'
      produces:
      - text/plain
      responses:
        "200":
          description: ""
      summary: Admin login
      tags:
      - Login
swagger: "2.0"